+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/payments/PutPendecyServiceRequest.kt
@@ -2,6 +2,7 @@ package br.com.dastec.gerenciasalao.controllers.requests.payments
 
 import br.com.dastec.gerenciasalao.validation.annotation.IsCurrency
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 import javax.validation.constraints.NotBlank
 
 
@@ -14,5 +15,5 @@ data class PutPendecyServiceRequest(
 
     @JsonAlias("value_payment")
     @IsCurrency
-    var valuePayment: Double,
+    var valuePayment: BigDecimal,
 )
