+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/pendency/PostAddPendencyRequest.kt
@@ -2,6 +2,7 @@ package br.com.dastec.gerenciasalao.controllers.requests.pendency
 
 import br.com.dastec.gerenciasalao.validation.annotation.IsCurrency
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 import javax.validation.constraints.NotNull
 
 data class PostAddPendencyRequest(
@@ -11,5 +12,5 @@ data class PostAddPendencyRequest(
 
     @JsonAlias("value_pendency")
     @IsCurrency
-    var valuePendency: Double
+    var valuePendency: BigDecimal
 )
