+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/services/PutServiceRequest.kt
@@ -2,6 +2,7 @@ package br.com.dastec.gerenciasalao.controllers.requests
 
 import br.com.dastec.gerenciasalao.validation.annotation.IsCurrency
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 import javax.validation.constraints.NotBlank
 
 data class PutServiceRequest(
@@ -15,5 +16,5 @@ data class PutServiceRequest(
     val idCategory: Long,
 
     @IsCurrency
-    val price: Double?
+    val price: BigDecimal?
 )
